Replace androidx-paging with multiplatform-paging
