--- extensions/android-paging3/android-test/build.gradle
@@ -1,27 +1,0 @@
-plugins {
-  alias(libs.plugins.android.library)
-  alias(libs.plugins.kotlin.android)
-  alias(libs.plugins.dokka)
-}
-
-android {
-  namespace 'app.cash.sqldelight.paging3'
-  compileSdk libs.versions.compileSdk.get() as int
-}
-
-dependencies {
-  testImplementation project(':drivers:sqlite-driver')
-  testImplementation project(':extensions:android-paging3')
-  testImplementation libs.androidx.paging3.runtime
-  testImplementation libs.androidx.recyclerView
-  testImplementation libs.truth
-  testImplementation libs.kotlin.test.junit
-  testImplementation libs.kotlin.coroutines.test
-}
-
-// workaround for https://youtrack.jetbrains.com/issue/KT-27059
-configurations.all {
-  resolutionStrategy.dependencySubstitution {
-    substitute module("${project.property("GROUP")}:runtime-jvm:${project.property("VERSION_NAME")}") with project(':runtime')
-  }
-}
