+++ extensions/android-paging3/build.gradle
@@ -1,27 +1,53 @@
 plugins {
-  alias(libs.plugins.kotlin.jvm)
+  id("org.jetbrains.kotlin.multiplatform")
   alias(libs.plugins.publish)
   alias(libs.plugins.dokka)
 }
 
 archivesBaseName = 'sqldelight-androidx-paging3'
 
-dependencies {
-  // workaround for https://youtrack.jetbrains.com/issue/KT-27059
-  api "${project.property("GROUP")}:runtime-jvm:${project.property("VERSION_NAME")}"
-  api libs.androidx.paging3.common
+kotlin {
+  ios()
+  jvm()
 
-  testImplementation project(':drivers:sqlite-driver')
-  testImplementation libs.truth
-  testImplementation libs.kotlin.test.junit
-  testImplementation libs.kotlin.coroutines.test
-}
+  sourceSets {
+    commonMain {
+      dependencies {
+        api project(':runtime')
+        api libs.app.cash.paging.common
+        implementation libs.kotlin.coroutines.core
+      }
+    }
+    commonTest {
+      dependencies {
+        implementation libs.kotlin.test
+        implementation libs.kotlin.coroutines.test
+        implementation libs.stately.concurrency
+      }
+    }
+    iosTest {
+      dependencies {
+        implementation project(':drivers:native-driver')
+        implementation libs.app.cash.paging.runtime
+      }
+    }
+    jvmTest {
+      dependencies {
+        implementation project(':drivers:sqlite-driver')
+        implementation libs.androidx.paging3.runtime
+      }
+    }
+  }
 
-// workaround for https://youtrack.jetbrains.com/issue/KT-27059
-configurations.all {
-  resolutionStrategy.dependencySubstitution {
-    substitute module("${project.property("GROUP")}:runtime-jvm:${project.property("VERSION_NAME")}") with project(':runtime')
+  configure([targets.iosX64, targets.iosArm64]) {
+    compilations.test {
+      kotlinOptions.freeCompilerArgs += ['-linker-options', '-lsqlite3 -L/usr/lib/x86_64-linux-gnu -L/usr/lib -L/usr/lib64']
+    }
   }
 }
 
-apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
+tasks.withType(org.jetbrains.kotlin.gradle.targets.native.tasks.KotlinNativeSimulatorTest) {
+  deviceId = "iPhone 14"
+}
+
+apply from: "$rootDir/gradle/gradle-mvn-push.gradle"
