+++ extensions/android-paging3/gradle.properties
@@ -1,4 +1,4 @@
 POM_ARTIFACT_ID=androidx-paging3-extensions
-POM_NAME=SQLDelight AndroidX Paging3 Coroutines Extension
-POM_DESCRIPTION=An implementation of AndroidX Paging3 backed by SQLDelight Query
+POM_NAME=SQLDelight AndroidX Paging3 Multiplatform Extension
+POM_DESCRIPTION=An implementation of AndroidX Paging3 Multiplatform backed by SQLDelight Query
 POM_PACKAGING=jar
