+++ extensions/android-paging3/src/commonMain/kotlin/app/cash/sqldelight/paging3/KeyedQueryPagingSource.kt
@@ -15,7 +15,11 @@
  */
 package app.cash.sqldelight.paging3
 
-import app.cash.paging.*
+import app.cash.paging.PagingSourceLoadParams
+import app.cash.paging.PagingSourceLoadResult
+import app.cash.paging.PagingSourceLoadResultError
+import app.cash.paging.PagingSourceLoadResultPage
+import app.cash.paging.PagingState
 import app.cash.sqldelight.Query
 import app.cash.sqldelight.Transacter
 import kotlinx.coroutines.withContext
