--- extensions/multiplatform-paging/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=multiplatform-paging-extensions
-POM_NAME=SQLDelight Multiplatform Paging Extension
-POM_DESCRIPTION=An implementation of Multiplatform AndroidX Paging backed by SQLDelight Query
-POM_PACKAGING=jar
