+++ settings.gradle
@@ -31,7 +31,6 @@ include ':drivers:sqljs-driver'
 include ':drivers:driver-test'
 include ':extensions:android-paging3'
 include ':extensions:android-paging3:android-test'
-include ':extensions:multiplatform-paging'
 include ':extensions:async-extensions'
 include ':extensions:coroutines-extensions'
 include ':extensions:rxjava2-extensions'
